Move sensor data functions from helper to connector
